Removed atomic_periods and atomic_groups from NNPInput.  Since these are just based on atomic number, moved the assignment into featurization, to minimize including additional things in NNPInput (since using slots now and no longer dynamic).
