+++ modelforge/utils/prop.py
@@ -49,8 +49,6 @@ class NNPInput:
         "per_atom_partial_charge",
         "box_vectors",
         "is_periodic",
-        "atomic_groups",
-        "atomic_periods",
     )
 
     def __init__(
@@ -72,209 +70,10 @@ class NNPInput:
         self.per_atom_partial_charge = per_atom_partial_charge
         self.box_vectors = box_vectors
         self.is_periodic = is_periodic
-        self._assign_atomic_group_period()
 
         # Validate inputs
         self._validate_inputs()
 
-    def _assign_atomic_group_period(self):
-        """
-        Assign atomic group to the input data.
-
-        Returns
-        -------
-        torch.Tensor
-            The atomic group tensor.
-        """
-        period = {
-            1: 1,
-            2: 1,
-            3: 2,
-            4: 2,
-            5: 2,
-            6: 2,
-            7: 2,
-            8: 2,
-            9: 2,
-            10: 2,
-            11: 3,
-            12: 3,
-            13: 3,
-            14: 3,
-            15: 3,
-            16: 3,
-            17: 3,
-            18: 3,
-            19: 4,
-            20: 4,
-            21: 4,
-            22: 4,
-            23: 4,
-            24: 4,
-            25: 4,
-            26: 4,
-            27: 4,
-            28: 4,
-            29: 4,
-            30: 4,
-            31: 4,
-            32: 4,
-            33: 4,
-            34: 4,
-            35: 4,
-            36: 4,
-            37: 5,
-            38: 5,
-            39: 5,
-            40: 5,
-            41: 5,
-            42: 5,
-            43: 5,
-            44: 5,
-            45: 5,
-            46: 5,
-            47: 5,
-            48: 5,
-            49: 5,
-            50: 5,
-            51: 5,
-            52: 5,
-            53: 5,
-            54: 5,
-            55: 6,
-            56: 6,
-            57: 6,
-            72: 6,
-            73: 6,
-            74: 6,
-            75: 6,
-            76: 6,
-            77: 6,
-            78: 6,
-            79: 6,
-            80: 6,
-            81: 6,
-            82: 6,
-            83: 6,
-            84: 6,
-            85: 6,
-            86: 6,
-            87: 7,
-            88: 7,
-            104: 7,
-            105: 7,
-            106: 7,
-            107: 7,
-            108: 7,
-            109: 7,
-            110: 7,
-            111: 7,
-            112: 7,
-            113: 7,
-            114: 7,
-            115: 7,
-            116: 7,
-            117: 7,
-            118: 7,
-        }
-        group = {
-            1: 1,
-            2: 18,
-            3: 1,
-            4: 2,
-            5: 13,
-            6: 14,
-            7: 15,
-            8: 16,
-            9: 17,
-            10: 18,
-            11: 1,
-            12: 2,
-            13: 13,
-            14: 14,
-            15: 15,
-            16: 16,
-            17: 17,
-            18: 18,
-            19: 1,
-            20: 2,
-            21: 3,
-            22: 4,
-            23: 5,
-            24: 6,
-            25: 7,
-            26: 8,
-            27: 9,
-            28: 10,
-            29: 11,
-            30: 12,
-            31: 13,
-            32: 14,
-            33: 15,
-            34: 16,
-            35: 17,
-            36: 18,
-            37: 1,
-            38: 2,
-            39: 3,
-            40: 4,
-            41: 5,
-            42: 6,
-            43: 7,
-            44: 8,
-            45: 9,
-            46: 10,
-            47: 11,
-            48: 12,
-            49: 13,
-            50: 14,
-            51: 15,
-            52: 16,
-            53: 17,
-            54: 18,
-            55: 1,
-            56: 2,
-            57: 3,
-            72: 4,
-            73: 5,
-            74: 6,
-            75: 7,
-            76: 8,
-            77: 9,
-            78: 10,
-            79: 11,
-            80: 12,
-            81: 13,
-            82: 14,
-            83: 15,
-            84: 16,
-            85: 17,
-            86: 18,
-            87: 1,
-            88: 2,
-            104: 4,
-            105: 5,
-            106: 6,
-            107: 7,
-            108: 8,
-            109: 9,
-            110: 10,
-            111: 11,
-            112: 12,
-            113: 13,
-            114: 14,
-            115: 15,
-            116: 16,
-            117: 17,
-            118: 18,
-        }
-
-        self.atomic_groups = torch.tensor(
-            [group[atomic_number.item()] for atomic_number in self.atomic_numbers]
-        )
-        self.atomic_periods = torch.tensor(
-            [period[atomic_number.item()] for atomic_number in self.atomic_numbers]
-        )
 
     def _validate_inputs(self):
         # Get shapes of the arrays
